fix: set WAF ACL body oversize handling
Update the WAF ACL rules with `body` blocks to set the oversize
request body size handling to `CONTINUE` with inspection.

This is required by AWS to continue to allow API management of
WAF ACLs.

Also updates the Terraform provider to the latest version to provide
support for the new `oversize_handling` attribute.